requirements.txt
django==1.8
djangorestframework==3.0
# django-rest-assured==0.2.0
# django-extensions==1.6.1
django-geojson==2.9.0
jsonfield==1.0.3
# factory-boy==2.7.0
# Shapely==1.5.15

# production - heroku
psycopg2
gunicorn
dj-database-url
